Drop polyfills from this library, 1.11 support
The `concat` polyfill is in Ember 1.13 so now built-in to Iverson.

The `hash` helper has a polyfill we can use directly in apps, and
additionally isn't reliant on the legacy `_registerHelper` API.

`hasBlock` is built-in to Ember 1.13 as a keyword, so basing it on
`template` via `reopen` is not needed.